Fixing CI errors
