+++ crates/pop-cli/src/common/mod.rs
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0
 
+pub mod builds;
 #[cfg(feature = "contract")]
 pub mod contracts;
 pub mod helpers;
